♻️ refactor: domain과 ui 로직 분리
