+++ src/views/InputView.js
@@ -1,26 +1,34 @@
-import readline from 'readline';
+import ReadLineAsync from './ReadLineAsync.js';
+import { Parser } from '../utils/Parser.js';
+import { Validator } from '../utils/Validator.js';
+import DEFINITION from '../constants/Definition.js';
+import MESSAGE from '../constants/Message.js';
 
 class InputView {
-  static readLineAsync(query) {
-    return new Promise((resolve, reject) => {
-      if (arguments.length !== 1) {
-        reject(new Error('arguments must be 1'));
+  static async inputName() {
+    while (true) {
+      try {
+        const inputName = await ReadLineAsync.readLineAsync(MESSAGE.INPUT.NAME);
+        const splittedName = Parser.splitName(inputName);
+        Validator.validateName(splittedName, DEFINITION.MAX_NAME_LENGTH);
+        return splittedName;
+      } catch (error) {
+        console.log(error.message);
       }
+    }
+  }
 
-      if (typeof query !== 'string') {
-        reject(new Error('query must be string'));
+  static async inputTryNumber() {
+    while (true) {
+      try {
+        const inputTryNumber = await ReadLineAsync.readLineAsync(MESSAGE.INPUT.TRY_NUMBER);
+        const parsedNumber = Number(inputTryNumber.trim());
+        Validator.validateTryNumber(parsedNumber, DEFINITION.MIN_GAME, DEFINITION.MAX_GAME);
+        return parsedNumber;
+      } catch (error) {
+        console.log(error.message);
       }
-
-      const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-      });
-
-      rl.question(query, input => {
-        rl.close();
-        resolve(input);
-      });
-    });
+    }
   }
 }
